chatbot/requirements.txt
fastapi
uvicorn
python-multipart
chromadb>=0.4.22
streamlit
langchain>=0.1.0
langchain-google-genai>=0.2.0
langchain-core>=0.2.0
pydantic>=2.0.0
google-generativeai>=0.3.0
langchain-community>=0.2.0
langchain-openai>=0.2.0
openai
google-generativeai
pdfplumber
python-docx
requests